requirements.txt
flask>=3.0
openai>=1.0
python-dotenv>=1.0
yfinance>=0.2
pandas>=2.0
numpy>=1.26
